requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
googletrans==4.0.0-rc1
streamlit==1.28.1
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
